0.3 Trait Mindedness System
Trait Mindedness System-
When the pawn enters the CC, the chamber will now decide whether the pawn has more traits towards being physical or more traits towards being mentally minded. This has been created so if a pawn gets a power then it will be a power that is more aligned with their thought process.