0.5 Beta
The system now has a working cycle from beginning to end! Playtesting can now begin.

-Creation Chamber
The chamber now has a button for the player to press for the process to work, along with this...The chamber now requires a Dnaide Canister and power going through the building for it to work.

-Powers
A bunch of powers have now been added to the system

-Super Speed is now Extreme Speed
-Extreme Precision
-Extreme Intelligence
-Extreme Understanding
-Extreme Dexterity
-Rock Skin
-Tree Skin

Balancing-
There have been a number of balance changes that have been made where certain powers will be more popular depending on the tier I think they belong. This will be subject to change as more powers are added and based on feedback of players.

Bugs-
Creation Chamber now plays a sound when the pawn leaves the chamber.